ag: Enable searching content of vim " register
Currently, only the word under the cursor can be searched easily via the
:Ag! command. This enables searching the contents of the unnamed
register. This is useful when you want to search for a multi-word
phrase without manually copy-pasting the term.

The intended use is to first yank the words to be searched for using
visual select, and then `y`. Then this command may be run.

Signed-off-by: Guru Das Srinagesh <gurooodas@gmail.com>